Implement jwt inside reset password service, Format the reset password (form file seperate), New hook file to verify password, movice common CSS from reset and login page
